+++ zencad/examples/1.GeomPrim/2.prim2d/widewire.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python3
+"""
+ZenCad API example: widewire
+"""
+
+from zencad import *
+
+zencad.lazy.fastdo = True
+zencad.lazy.encache = False
+zencad.lazy.decache = False
+
+wr = wire_builder()
+wr.line(10,10)
+wr.line(20,20)
+wr.line(30,0)
+wr.interpolate([
+	[40, -20],
+	[30, -40]
+])
+wr.interpolate([
+	[50, -60],
+	[60, -110]
+])
+
+w = widewire(wr.doit(), 5)
+
+disp(wr.doit(), zencad.color.red)
+disp(w)
+show()