+++ zencad/geom/ops1d2d.py
@@ -2,7 +2,14 @@ import pyservoce
 import evalcache
 
 from zencad.lazifier import lazy, shape_generator, nocached_shape_generator
-from zencad.util import points, vectors
+from zencad.util import points, vectors, point3, vector3
+
+from zencad.geom.prim1d import segment, vertex
+from zencad.geom.prim2d import circle
+from zencad.geom.ops3d import pipe
+from zencad.geom.boolean import union
+
+import math
 
 
 @lazy.lazy(cls=shape_generator)
@@ -114,3 +121,59 @@ def chamfer2d(shp, r, refs=None):
 @lazy.lazy(cls=shape_generator)
 def fix_face(f):
     return pyservoce.fix_face(f)
+
+
+def _wideedge(spine, rad, last_p0, last_p1, circled_joints):
+    # TODO: переимплементировать без сегмента
+    import zencad
+
+    if spine.shapetype() != "edge":
+        raise Exception("argument 'curve' should be edge")  
+
+    ad0 = spine.endpoints()[0]
+    ad1 = spine.endpoints()[1]
+
+    d10 = spine.d1(spine.range()[0]).cross(vector3(0,0,1)).normalize()
+    d11 = spine.d1(spine.range()[1]).cross(vector3(0,0,1)).normalize()
+    #p0 = pyservoce.vertex((pt*rad/2).to_point3())
+    #p1 = pyservoce.vertex((-pt*rad/2).to_point3())
+
+    # начальные точки
+    p00 = ad0 + (d10 * rad)
+    p01 = ad0 - (d10 * rad) 
+
+    #конечные точки
+    p10 = ad1 + (d11 * rad)
+    p11 = ad1 - (d11 * rad)
+
+    perp = pyservoce.segment(p00, p01)
+    wc = pyservoce.pipe(profile=perp, spine=spine, mode="corrected_frenet")
+
+    wc = wc.faces()[0]
+
+    if circled_joints is False:
+       if last_p1 is not None and last_p0 is not None:
+            wc = wc + pyservoce.polygon([last_p0, p00, ad0])
+            wc = wc + pyservoce.polygon([last_p1, p01, ad0])
+    else:
+       if last_p1 is not None and last_p0 is not None:
+            wc += pyservoce.circle(r=rad).mov(vector3(ad0))
+
+    return wc, p10, p11 #union(edgs) 
+
+@lazy.lazy(cls=shape_generator)
+def widewire(spine, r, circled_joints=True, circled_ends=True):
+    edges = spine.edges()
+    p0 = None
+    p1 = None
+    rad = r
+    arr =[]
+    for e in edges:
+        f, p0, p1 = _wideedge(e, rad, p0, p1, circled_joints=circled_joints)
+        arr.append(f)
+
+    if circled_ends:
+        arr.append(pyservoce.circle(r=rad).move(spine.endpoints()[0]))
+        arr.append(pyservoce.circle(r=rad).move(spine.endpoints()[1]))
+
+    return pyservoce.unify(union(arr))